LayoutTests/media/modern-media-controls/tracks-support/hidden-tracks-expected.txt

This test checks that the tracks contextmenu doesn't show any track as selected if all the text tracks for a media element are marked as "hidden".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
Setting mode to 'hidden' for all text tracks via JavaScript...
Tapping tracks button...
PASS checkedItems.length is 1
PASS checkedItems[0].title is "Off"
PASS successfullyParsed is true

TEST COMPLETE
